+++ my-tiny-spring/src/main/java/com/yrb/tinyioc/aop/Pointcut.java
@@ -1,0 +1,14 @@
+package com.yrb.tinyioc.aop;
+
+/**
+ * @author bjyangrubing
+ * @createTime 2017/8/8 11:12
+ * Description: 
+ */
+public interface Pointcut
+{
+	ClassFilter getClassFilter();
+
+	MethodMatcher getMethodMatcher();
+
+}