+++ my-tiny-spring/src/main/java/com/yrb/tinyioc/aop/PointcutAdvisor.java
@@ -1,0 +1,11 @@
+package com.yrb.tinyioc.aop;
+
+/**
+ * @author bjyangrubing
+ * @createTime 2017/8/8 11:17
+ * Description: 
+ */
+public interface PointcutAdvisor extends Advisor
+{
+	Pointcut getPointcut();
+}