+++ my-tiny-spring/src/main/java/com/yrb/tinyioc/beans/factory/AbstractBeanFactory.java
@@ -1,6 +1,7 @@
 package com.yrb.tinyioc.beans.factory;
 
 import com.yrb.tinyioc.beans.BeanDefinition;
+import com.yrb.tinyioc.beans.BeanPostProcessor;
 
 import java.util.ArrayList;
 import java.util.Iterator;
@@ -19,6 +20,8 @@ public abstract class AbstractBeanFactory implements BeanFactory
 
 	private final List<String> beanDefinitionNames = new ArrayList<String>();
 
+	private List<BeanPostProcessor> beanPostProcessors = new ArrayList<BeanPostProcessor>();
+
 	@Override
 	public Object getBean(String name) throws Exception
 	{
@@ -31,8 +34,23 @@ public abstract class AbstractBeanFactory implements BeanFactory
 		if (bean == null)
 		{
 			bean = doCreateBean(beanDefinition);
+			initializeBean(bean, name);
 		}
 		return bean;
+	}
+
+	protected void initializeBean(Object bean, String name) throws Exception
+	{
+		for (BeanPostProcessor beanPostProcessor : beanPostProcessors)
+		{
+			bean = beanPostProcessor.postProcessBeforeInitialization(bean, name);
+		}
+
+		//TODO :call initialize method
+		for (BeanPostProcessor beanPostProcessor : beanPostProcessors)
+		{
+			bean = beanPostProcessor.postProcessAfterInitialization(bean, name);
+		}
 	}
 
 	public void registerBeanDefinition(String name, BeanDefinition beanDefinition) throws Exception
@@ -50,10 +68,45 @@ public abstract class AbstractBeanFactory implements BeanFactory
 		}
 	}
 
+	protected Object doCreateBean(BeanDefinition beanDefinition) throws Exception
+	{
+		Object bean = createBeanInstance(beanDefinition);
+		beanDefinition.setBean(bean);
+		applyPropertyValues(bean, beanDefinition);
+		return bean;
+	}
+
+	protected abstract void applyPropertyValues(Object bean, BeanDefinition beanDefinition) throws Exception;
+
 	/**
-	 * 根据beanDefinition, 初始化bean
+	 * 通过Class对象来创造一个对象
 	 * @param beanDefinition
 	 * @return
+	 * @throws IllegalAccessException
+	 * @throws InstantiationException
 	 */
-	protected abstract Object doCreateBean(BeanDefinition beanDefinition) throws Exception;
+	private Object createBeanInstance(BeanDefinition beanDefinition)
+			throws IllegalAccessException, InstantiationException
+	{
+		return beanDefinition.getBeanClass().newInstance();
+	}
+
+	public void addBeanPostProcessor(BeanPostProcessor beanPostProcessor) throws Exception
+	{
+		this.beanPostProcessors.add(beanPostProcessor);
+	}
+
+	public List<Object> getBeansForType(Class type) throws Exception
+	{
+		List<Object> beans = new ArrayList<Object>();
+		for (String beanDefinitionName : beanDefinitionNames)
+		{
+			if (type.isAssignableFrom(beanDefinitionMap.get(beanDefinitionName).getBeanClass()))
+			{
+				beans.add(getBean(beanDefinitionName));
+			}
+		}
+		return beans;
+	}
+
 }