+++ my-tiny-spring/CHANGELOG.md
@@ -17,3 +17,5 @@
  - ApplicationContext登场
 ## 7. step-7-method-interceptor-by-jdk-dynamic-proxy
  - 使用JDK动态代理实现AOP织入
+## 8. step-8-invite-pointcut-and-aspectj
+ - 使用Aspectj管理切面,并引入BeanPostProcessor